Frontend-offline: Added UI for kilometers and included it into pdf

+++ mobile-offline/components/mechanic/library/forms/RepairForm.tsx
@@ -93,7 +93,7 @@ export default function RepairForm({ repair, setRepair }: Props) {
     const repairData: RepairData = {
       uuid: repair?.uuid || "",
       type: type,
-      kilometers: parseInt(kilometers) || null,
+      kilometers: parseFloat(kilometers) || null,
       price: parseFloat(price) || null,
       repairDate: date,
       options: options,
